requirements.txt
aiogram==3.20.0.post0
python-socketio[client]==5.11.2
aiosqlite==0.20.0
beautifulsoup4==4.12.3
requests==2.31.0
